Move nvidia-settings usage to separate thread WIP
Some experimentation with moving the querying of nvidia-settings to a separate thread and detaching it. nvidia-settings called from within the main function causes huge delays, and by detaching a thread and writing this to a separate structure, the values can be updated albeit not very fast.

Will experiment some more with this.